Extract query related tests into queryTests.test.ts
